fix: Move makeColumns to columns.go as it was unit tested in columns_test.go
